backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
signxml
lxml
jinja2
